Stores refactor
