+++ template/src/components/views/stores/Stores.js
@@ -4,13 +4,21 @@ import Divider from '@material-ui/core/Divider';
 import { AggregationTypes } from 'lib/sdk';
 import { FormulaWidget, CategoryWidget } from 'components/common/widgets';
 import { LayerStyle } from 'components/layers/StoresLayer';
-import { setViewState, addLayer } from 'config/cartoSlice';
+import {
+  setViewState,
+  addLayer,
+  addSource,
+  removeLayer,
+  removeSource,
+} from 'config/cartoSlice';
 import { currencyFormatter } from 'utils/numberFormatters';
+import { SOURCE_ID, LAYER_ID } from './common';
 
-export default function StoreList() {
+export default function Stores() {
   const dispatch = useDispatch();
 
   useEffect(() => {
+    // Change zoom
     dispatch(
       setViewState({
         latitude: 31.802892,
@@ -19,16 +27,29 @@ export default function StoreList() {
         transitionDuration: 500,
       })
     );
+    // Add stores source
     dispatch(
-      addLayer({ id: 'storesLayer', source: 'storesSource', selectedStore: null })
+      addSource({
+        id: SOURCE_ID,
+        data:
+          'SELECT store_id, zip, storetype, state, revenue, the_geom_webmercator FROM mcdonalds',
+      })
     );
+    // Add layer
+    dispatch(addLayer({ id: LAYER_ID, source: SOURCE_ID, selectedStore: null }));
+
+    // Clean up when leave
+    return function cleanup() {
+      dispatch(removeLayer(LAYER_ID));
+      dispatch(removeSource(SOURCE_ID));
+    };
   }, [dispatch]);
 
   return (
     <div>
       <FormulaWidget
         title='Total revenue'
-        data-source='storesSource'
+        data-source={SOURCE_ID}
         operation-column='revenue'
         operation={AggregationTypes.SUM}
         formatter={currencyFormatter}
@@ -37,7 +58,7 @@ export default function StoreList() {
       <Divider />
       <CategoryWidget
         title='Revenue by store type'
-        data-source='storesSource'
+        data-source={SOURCE_ID}
         column='storetype'
         operation-column='revenue'
         operation={AggregationTypes.SUM}
