+++ template/src/components/views/stores/common.js
@@ -1,0 +1,2 @@
+export const SOURCE_ID = 'storesSource';
+export const LAYER_ID = 'storesLayer';
