+++ template/src/routes.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import Home from 'components/views/Home';
-import Stores from 'components/views/Stores';
-import StoreList from 'components/views/StoreList';
-import StoreDetail from 'components/views/StoreDetail';
+import Stores from 'components/views/stores/Stores';
+import StoresDetail from 'components/views/stores/StoresDetail';
 import Kpi from 'components/views/Kpi';
 
 const routes = [
@@ -10,14 +9,8 @@ const routes = [
     path: '/',
     element: <Home />,
     children: [
-      {
-        path: '/stores',
-        element: <Stores />,
-        children: [
-          { path: '/', element: <StoreList /> },
-          { path: ':id', element: <StoreDetail /> },
-        ],
-      },
+      { path: '/stores', element: <Stores /> },
+      { path: '/stores/:id', element: <StoresDetail /> },
       { path: '/kpi', element: <Kpi /> },
     ],
   },
